+++ components/core/markdown/src/main/java/com/flipperdevices/core/markdown/ComposableMarkdown.kt
@@ -37,7 +37,7 @@ fun ComposableMarkdown(
 
 @Composable
 fun markdownColors(
-    backgroundCode: Color = LocalPallet.current.text20,
+    backgroundCode: Color = LocalPallet.current.text8,
     text: Color = LocalPallet.current.text100,
     link: Color = LocalPallet.current.accentSecond,
     dividerColor: Color = LocalPallet.current.divider12
@@ -48,7 +48,7 @@ fun markdownColors(
     codeBackground = backgroundCode,
     inlineCodeBackground = backgroundCode,
     dividerColor = dividerColor,
-    inlineCodeText = backgroundCode
+    inlineCodeText = text
 )
 
 @Composable