+++ components/infrared/editor/src/main/kotlin/com/flipperdevices/infrared/editor/compose/screen/ComposableInfraredEditorScreenReady.kt
@@ -2,9 +2,11 @@ package com.flipperdevices.infrared.editor.compose.screen
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.lazy.LazyColumn
 import androidx.compose.foundation.lazy.itemsIndexed
 import androidx.compose.runtime.Composable
@@ -89,6 +91,7 @@ internal fun ComposableInfraredEditorScreenReady(
                     )
                 }
             }
+            item { Box(modifier = Modifier.navigationBarsPadding()) }
         }
     }
 }